Fuzz combinations of two modules (#7144)
The new Two fuzz testcase handler generates two wasm files and
then runs them in V8, linking them using JS. It then optimizes at least one of
the two and runs it again, and checks for differences.

This is similar to the Split fuzzer for wasm-split, but the two wasm files
are arbitrary and not the result of splitting.

Also lower CtorEval priority a little. It is not very important.