[NFC] Encode reference types with bit packing (#7142)
Value types were previously represented internally as either enum values
for "basic," i.e. non-reference, non-tuple types or pointers to
`TypeInfo` structs encoding either references or tuples. Update the
representation of reference types to use one bit to encode nullability
and the rest of the bits to encode the referenced heap type. This allows
canonical reference types to be created with a single logical or rather
than by taking a lock on a global type store and doing a hash map lookup
to canonicalize.

This change is a massive performance improvement and dramatically
improves how performance scales with threads because the removed lock
was highly contended. Even with a single core, the performance of an O3
optimization pipeline on a WasmGC module improves by 6%. With 8 cores,
the improvement increases to 29% and with all 128 threads on my machine,
the improvement reaches 46%.

The full new encoding of types is as follows:

 - If the type ID is within the range of the basic types, the type is
   the corresponding basic type.

 - Otherwise, if bit 0 is set, the type is a tuple and the rest of the
   bits are a canonical pointer to the tuple.

 - Otherwise, the type is a reference type. Bit 1 determines the
   nullability and the rest of the bits encode the heap type.

Also update the encodings of basic heap types so they no longer use the
low two bits to avoid conflicts with the use of those bits in the
encoding of types.