Update flake8 and fix errors
The flake8 we were running on CI was too old and began giving spurious
errors about the uninterpreted contents of f-strings. Update to the
latest flake8 and fix all the new errors, including the previously
incorrect comment syntax in the .flake8 file.

Also remove scripts/storage.py, since it didn't seem to be used for
anything we currently need.