Fix UBSan on CI (#7173)
The UBSan builder started failing with an error about a misaligned store
in wasm-ctor-eval.cpp. The store was already done via `memcpy` to avoid
alignment issues, but apparently this is no longer enough. Use `void*`
as the destination type to further avoid giving the impression of
guaranteed alignment.

Also fix UB when executing std::abs on minimum negative integers in
literal.cpp.