ClusterFuzz: Add a --no_retry option for local debugging (#7187)
When we run ClusterFuzz's run.py locally, we need to investigate any
errors in creating fuzz testcases (as those are real bugs in wasm-opt).
Previously we just checked for a "retry" message in the output of run.py
(as the script retries automatically - we don't want a wasm-opt bug to
stop ClusterFuzz), but that makes it impossible to debug the problem,
since the retry tramples the data. The new option, used only locally,
stops at the error so it can be investigated.