Fuzzing: Handle instantiation errors in ClusterFuzz (#7166)
If the module has an invalid segment offset, for example, it will error. We
should not error in the fuzzer on such rare cases.

Also add logging of the actual error, which matches what we do for
errors elsewhere, and makes debugging easier. As a result another place
needs to look for the prefix now, and not the entire string (since we
append the error contents).