Model constraints on StringNew children more precisely (#7193)
The stringref spec has a bug where the string.new... instructions that
take array reference parameters do not have the proper type annotations,
which means IRBuilder does not have a single type that it knows the
array reference must have. We previously fudged this by using `arrayref`
as the type constraint, but the fuzzer was able to generate a valid test
case that was parsed into invalid IR due to this imprecision.

Fix the bug by adding special constraint types for "any i8 array
reference" and "any i16 array reference" just for use with the offending
string.new instructions.

Fixes #7191.