Fuzzer: Fix up rethrows after changes (#7207)
We may move a rethrow into an invalid position when we mutate() etc. This finds
such invalid rethrows and replaces them with something trivial so that we validate.

To do so, use an expression stack rather than a control flow stack in the fixer,
as we need to see which child of a try the rethrow ends up being (so we need the
full path from it to the try).