Fix typing of struct RMW ops with null references
When the struct reference is null, there is no struct type from which to
look up the proper field type when finalizing a struct RMW or cmpxchg
expression. We previously reused code for this case from StructGet, but
unlike StructGet, the RMW operations have additional operands that can
constrain the type of the retrieved struct field. Take these operands
into account when computing a type for the RMW expressions.