[NFC] Fix note about manually generated test output (#7221)
The note mentioned update_lit_checks.py in the first line, which made
update_lit_checks.py think that it _should_ update the test output, even
though the comment was explicitly about how it should not update the
test output. Reword the comment to avoid the problem.

Also make the check in the update script stricter for good measure.