Update Heap2Local for struct RMW and cmpxchg (#7222)
Update the escape analysis to note that RMW modification values and
cmpxchg replacement values can escape, but other operands do not escape.
When the ref operands are being replaced with locals, lower the RMW and
cmpxchg ops to the corresponding local and binary operations, being
careful to use scratch locals necessary to prevent reordering problems.

Add tests in a new file that will be ignored by the fuzzer until we have
better fuzzing support for RMW and cmpxchg ops.