Add V8 testing on CI (#7208)
This adds v8 from jsvu to CI, and a single lit test that uses it, to get started.

The tests are skipped on Alpine Linux, which jsvu is not compatible with. To
do so, we just delete the tests before running lit (for lack of a better
skipping mechanism, as `lit --xfail` is not supported in our lit on CI).