Do not optimize tag types in SignaturePruning
Now that we preserve tag heap types in the IR, it was possible for
SignaturePruning to optimize heap types used by tags in such a way that
the tag uses would no longer be valid. An ideal fix would be to have
SignaturePruning analyze and optimize tag usage as well as calls, but
for now just skip optimizing any heap type used in a tag.