Apply warning and error compile flags only to 1P code
Move the unit test source under src/ so that all of our 1P code and none
of our 3P code is under src/, then move the configuration of all the
error and warning flags to src/CMakeLists.txt so it only applies to 1P
code. This will make it easier to add 3P code that does not compile
cleanly with all the warnings and errors we enable for our own code.