Update ConstantFieldPropagation for struct RMW ops (#7263)
RMW operations generally modify the field they access, so they inhibit
constant field propagation. The exceptions are xchg and cmpxchg
operations, which are like sets (conditional sets in the case of
cmpxchg) that happen to do gets as well. Just like for sets, we can
still optimize when the value written by xchg or cmpxchg happens to be
either the same constant value we know is already in the field or is a
copy from some instance of the same field.

There is potential for further optimization when we know an xchg or
cmpxchg sets a value that was read from some other field that we know
has the same constant value, but that's a missing optimization for plain
sets as well. There is also further optimization potential when we can
reason that a cmpxchg that would write a different value will never
perform the write because the comparison would never succeed. Add tests
with TODOs so we remember these potential optimizations in the future.