Fix optimization of seqcst RMWs in OptimizeInstructions
We previously reasoned that RMWs that do not change in-memory values
could be optimized to just atomic gets because they did not necessarily
synchronize with subsequent atomic reads of the same location. This is
not correct for seqcst RMWs, though, since subsequent seqcst reads in
the total order of seqcst operations are _required_ to read from and
synchronize with these RMWs, even though they do not change the
in-memory values. Fix the optimization to only work on acquire-release
RMWs and RMWs to unshared memory.