Do not emit fences when peephole optimizing struct RMWs (#7281)
We previously reasoned that we had to emit fences when optimizing
sequentially consistent RMWs to unshared memory to preserve the RMWs'
effect on the total order of sequentially consistent operations.
However, that total order alone is insufficient to force any read event
to observe some other write event; that requires a synchronization edge
to be formed via an atomic write and read pair. RMWs to unshared memory
can never be part of such a synchronization edge, so removing them
cannot introduce new behaviors. Improve our optimization to stop
emitting the unnecessary fences.