Fuzzer: Simplify and improve emitting of unreachable code (#7295)
In the earlier days of the fuzzer we would call makeUnary with unreachable
sometimes, and then emit an unreachable i32.eqz for example. Later we
added mutate() which randomly emits unreachable into places, which gives
even better coverage - no need to manually handle unreachability in every
single instruction. This PR removes the old form.

Also remove the "Important" label from makeUnreachable. The
unreachable instruction is now one of many that have that type, and we
have no reason to prefer it to others.

These changes have the effect of improving the chance to emit interesting
instructions with unreachable type, like throw, throw_ref.