[EH] Fuzz catch+rethrow on the JS side (#7287)
Add a parameter to call-export, call-ref, where the first bit says "catch and
rethrow any exception". This has no observable effect in binaryen, but causes us
to use different code paths in VMs (for example, a wasm exception may end up
caught in JS, then thrown in JS; passing wasm exnrefs to JS for throwing is not
possible otherwise, so this is the closest we can get).