[EH] Fuzzer: Fix infinite loop with nested exnrefs (#7294)
When we need an exnref value we may create a try-catch with a throw (so
when we catch it, we get an exnref). But if the exception tag contains an
exnref, then we will try to create another exnref in the throw, leading to
recursion (possibly infinite). To avoid this, just create a trivial tag with no
values when we just want a trivial value.