Fuzzer: Adjust parameters by input size, sometimes (#7276)
Rather than always have a fixed max of heap types, globals, etc. allow more if
the input size is large. This adds variety.

There is a cost to this variety, we go from 0.89% ignored runs to 1.1% (that is,
slightly more runs contain something that prevents us from doing the work
we want, like hitting a host limitation or seeing the testcase just traps). This
increase seems small enough to be reasonable, and we only do this half the
time, so the old behavior is still being tested at half throughput.