[NFC] Store function effects on functions (#7302)
This fixes an unsettling bug where, with a long-enough sequence of passes,
we might create a function A, compute it has no effects, remove that
function, create another function with the same but with effects, and think
the old effects apply to it. This is basically a danger of using a map from
function name to effects.

To avoid that, store the effects on function objects themselves.