Fuzzer: Fix handling of JS null exceptions (#7303)
We have a test

try {
  ..
} catch (e) {
  e.a; // must be valid
}

That test will catch various wasm exception issues, but it turns out
that it also fails on one specific JS exception: null. It is ok to do `e.a` on
numbers, objects, anything really, all except for null. This PR adds a
guard for that.