Fuzzer: Avoid massive initial sizes for memories (#7305)
We generate random segments and then make the memory's initial
size big enough to accomodate them, but if the size is massive then
we will just OOM anyhow (or even not validate in wasm32 in some
cases). To avoid that, put a limit on the maximum initial memory
size as influenced by segments.