Fuzzer: Add an option to preserve imports and exports (#7300)
Normally the fuzzer will add new imports (for the JS stuff we can
call), and new exports as it adds functions. This adds an option to
avoid that, and instead keep the imports and exports fixed. This
is useful when we are used to modify an existing fuzzer's testcase,
as its connection to the JS side should be considered fixed (and it
will run in that fuzzer's JS, not ours).

This is added as

--fuzz-preserve-imports-exports

for wasm-opt.