[NFC] Update a comment about synchronization in GUFA
The comment mentioned that in principle we might be able to optimize
atomic struct.gets when we know what the stored value we are loading
will be. This is not correct because the write and the read would still
create a synchronization edge in that case. Update the comment
accordingly.