Implement exact reference types
Introduced in the Custom RTTs proposal, exact reference types are
referencdes to a particular heap type and not any of its subtypes. The
exception is that exact references to the bottom type of each heap type
hierarchy are subtypes of exact references to any other type in the
hierarchy. This maintains the property that each hierarchy of reference
types is a lattice.

Implement, construction, equality, hashing, subtyping, LUB, and GLB
operations for exact reference types, but do not yet use them in the IR
or implement parsing for them.