Fuzzer: Modify functions late (#7313)
Previously we would create a function and then immediately modify it in
interesting ways (replace code, move code around, etc.). We also modified any
initial functions (in the content we are given to build on top of) at the very
start. That was simple, by modifying later we can get some benefits. This PR
makes us work in a loop:

while more:
  if random:
    add function
  else
    modify some random function previously added

This lets us unify the modification code to one place (rather than have it
separate for initial functions and for created functions). Also, by modifying
late, we can end up with calls from early functions to late ones, as we may
create A and B before modifying A, and end up mutating something into a call to
B. Previously we never had such forward calls.

For such late modding to work, the FunctionCreationContext needs to be a little
smarter, since we don't use a single context for creation and modification - now
the modification may happen later. Specifically, it needs to scan the body to
see the next label index that is valid to add, which we did not need before.
This PR also moves some things out of the context destructor, because now we
will run it more than once on the same function, and some things only need to
happen once like adding the hang limit checks.