[ci] Update Ubuntu 20.04 ci job to Ubuntu 22.04 (#7326)
If you pick a recent workflow run (here is the latest run on main
https://github.com/WebAssembly/binaryen/actions/runs/13530775040 as of
writing this PR) you'll see following at the bottom

![image](https://github.com/user-attachments/assets/81f93f61-28f8-4346-9b8f-f50b239457bf)

If you click of the link for this warning you'll be taken here
https://github.com/actions/runner-images/issues/11101 . It states that
the Ubuntu 20.04 runner was deprecated at the start of this month, and
will be fully unsupported by the 1st April . Therefore in my opinion is
worth removing the Ubuntu 20.04 job from the ci rather than risk trying
to run on an unsupported runner. If you need to keep supporting this
version of Ubuntu then a job could be created in a separate PR, which
builds inside a Docker container with a Ubuntu 20.04 base image.