[NFC] Avoid huge output in validator messages (#7334)
Nested expressions that all fail to validate can lead to quadratic output
and OOMs. Instead, put a limit on the size we print, and after it only
emit truncated error messages of fixed size.

Also make some of the types more precise in this code, which makes
it easier to read.

Fixes #7333