Parsing and emitting of exact types (#7342)
Implement text and binary parsing of exact references types, including
parsing of reference type shorthands with the exact prefix. Also
implement binary writing of exact types. When writing exact types when
custom descriptors are not enabled, generalize the types to their
inexact versions. This is very similar to how we generalize function
types when GC is not enabled, for instance.