Handle exact references when fixing non-defaultable locals (#7357)
Some passes leave non-defaultable locals in an invalid state where not
all the gets of the local are structurally dominated by sets. To fix
this, the pass runner automatically calls
`TypeUpdating::handleNonDefaultableLocals` to find the problematic
locals and make them nullable. This function did not previously preserve
exactness in the updated local types. Fix it and add a test.