[Wasm GC] Fix GUFA on trampled params in TrapsNeverHappens mode (#7368)
GUFA in TNH mode will infer that this will trap:

(func $foo (param $x)
  (ref.cast $T (local.get $x))
)

(func $bar
  (call $foo (X))
)

If X's type will cause a trap when cast to T, then we infer that the call
will trap, and optimize. However, we were missing a check for the
param being trampled,

(func $foo (param $x)
  (local.set $x (Y))     ;; !!!!!!!!!!
  (ref.cast $T (local.get $x))
)

Then if Y can be cast, we should not actually trap.

Fix this by just tracking which params are written to. We only look
in the first basic block anyhow, and traverse it in order, so that is
enough.

Fixes #7366