Fix clang-format on CI (#7374)
Make it print the diff again. Previously, when there was a formatting
problem, the script would exit the first time it ran `git clang-format`
without printing the results. Update the script to ignore errors on that
first run to ensure it makes it to the part where the results are
printed.